+++ gpuScheduler/multicriteria/multicriteria.hpp
@@ -8,7 +8,7 @@
 #include "../json.hpp"
 
 typedef std::string VariablesType;
-typedef double WeightType;
+typedef float WeightType;
 
 class Multicriteria {
 public:
